removed unneccessary comments and consoles. added named exports
