+++ src/footer.js
@@ -1,14 +1,12 @@
 // Footer components
-const Footer = () => {
+export const Footer = () => {
   const hour = new Date().getHours();
   const openingHour = 10;
   const closingHour = 22;
   const isOpen = hour >= openingHour && hour <= closingHour;
-  console.log(isOpen);
 
   return (
     <footer className="footer">
-      {/* Conditional Redenring - only render footer text if isOpen is true (Using ternary operator) */}
       {isOpen ? (
         <div className="order">
           <p>We're open until {closingHour}:00. Come visit us</p>
@@ -24,5 +22,3 @@ const Footer = () => {
     </footer>
   );
 };
-
-export default Footer;
