+++ src/header.js
@@ -1,10 +1,8 @@
 // Header Component
-const Header = () => {
+export const Header = () => {
   return (
     <header className="header">
       <h1>Fast React Pizza Co.</h1>
     </header>
   );
 };
-
-export default Header;
